Refresh DB objects after calling locked PiSi methods
Any method in the PiSi API with the decorator `@locked` invalidates all of our LazyDB objects (`self.packagedb` and the like). Any time we call one of those, we need to recreate all of our LazyDB objects (like `PackageKitPisiBackend.get_db` does) before the next time we query a database.