backends/eopkg: Don't error if resolve can't find anything
Resolve is meant to return empty if it can't resolve anything. To do
this we no longer set the ERROR_PACKAGE_NOT_FOUND error in __get_package()
and instead raise a PkError that can be caught by resolve.

This fixes strange discover behaviour where it would try and resolve pkgs
only found in the flatpak appstream file against packagekit then stop resolving
when it encountered an error.

A follow up patch will be needed to double check the behaviour of other
functions and whether they should error if the package can't be found.